setup to store data in json file and few other changes
In phoneScan I added changes to allow the data that is presented back onto the user to be stored in a data.json file. Additionally added a button to the GUI that allows user to end the main activity peacefully. Also modified  the create window method such that the the windows that are created bigger and added a title to the new windows that are created when button is pressed